+++ frontend/dev-dist/sw.js
@@ -81,7 +81,7 @@ define(['./workbox-d3a36d94'], (function (workbox) { 'use strict';
     "revision": "3ca0b8505b4bec776b69afdba2768812"
   }, {
     "url": "/index.html",
-    "revision": "0.mnho850hn5"
+    "revision": "0.urj3gtm30e8"
   }], {});
   workbox.cleanupOutdatedCaches();
   workbox.registerRoute(new workbox.NavigationRoute(workbox.createHandlerBoundToURL("/index.html"), {